+++ Convertidor/Dtos/Presupuesto/PrePucCompromisosUpdateDto.cs
@@ -10,9 +10,9 @@
         public int CodigoPuc { get; set; }
         public int FinanciadoId { get; set; }
         public int CodigoFinanciado { get; set; }
-        public int Monto { get; set; }
-        public int MontoCausado { get; set; }
-        public int MontoAnulado { get; set; }
+        public decimal Monto { get; set; }
+        public decimal MontoCausado { get; set; }
+        public decimal MontoAnulado { get; set; }
         public string Extra1 { get; set; } = string.Empty;
         public string Extra2 { get; set; } = string.Empty;
         public string Extra3 { get; set; } = string.Empty;
